requirements.txt
asgiref
asttokens==2.0.8
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.1
Django==4.1
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.6
executing==1.1.0
fonttools==4.37.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.6
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.1.0
python-dotenv==1.0.0
pytz==2023.3
PyX==0.15
PyYAML==6.0
requests==2.31.0
scapy==2.4.5
scipy==1.9.3
simplejson==3.19.1
sqlparse
stack-data==0.5.1
traitlets==5.4.0
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.5
